remove pco2 related parts, QtGui change to QtWidgets Close pyqtwidgets when closing the app fix localtest intensities (to float)

+++ pHox.py
@@ -76,7 +76,7 @@ class Spectro_localtest(object):
 
     async def get_intensities(self, num_avg=1, correct=True):
         def _get_intensities():
-            sp = self.test_df["0"].values + random.randrange(-1000, 1000, 1)
+            sp = self.test_df["0"].astype('float').values + random.randrange(-1000, 1000, 1)
             return sp
 
         while self.busy:
@@ -88,7 +88,7 @@ class Spectro_localtest(object):
         return sp
 
     def get_intensities_slow(self, num_avg=1, correct=True):
-        sp = self.test_df["0"].values + random.randrange(-1000, 1000, 1)
+        sp = self.test_df["0"].astype('float').values + random.randrange(-1000, 1000, 1)
         return sp
 
 class Spectro_seabreeze(object):
@@ -342,7 +342,7 @@ class Common_instrument(object):
         
         if nAver < 3:
             logging.error(' num of Volt measurements: {}'.format(str(nAver)))
-            
+
         return Voltage
 
     def calc_wavelengths(self):
